requirements.txt
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
easyocr>=1.7.1
spacy>=3.7.2
transformers>=4.37.0
torch>=2.2.0
pillow>=10.2.0
numpy>=1.26.0
pandas>=2.2.0
opencv-python>=4.9.0
pdf2image>=1.17.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
requests>=2.31.0
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.1.0
flake8>=7.0.0
httpx>=0.26.0
